requirements.txt
fastapi==0.97.0
PyJWT==2.7.0
uvicorn==0.22.0

